修复CD部署时间

+++ models/cd_deployment_logs.go
@@ -47,7 +47,7 @@ type NewDeploymentArray struct {
 	Namespace           string
 	Match_tag           string //是否匹配规则
 	Start_time          time.Time
-	MinReadySeconds     int8
+	MinReadySeconds     int32
 }
 
 func (cd *CDDeploymentLogs) TableName() string {
